+++ NEWS.md
@@ -3,6 +3,8 @@
 * Fix bug in `t_rsp` of not showing p-vlaue, odds ratio and CIs when `strata_data` is not `NULL`
 * Add stratified analysis for `t_forest_rsp` and `t_forest_tte`, stratified analysis is footnoted in `g_forest`.
 * Addes `footnotes`, `footnotes<-` and `add_footnotes<-` functions to deal with footnotes
+* Add argument `conf_int` for confidence interval level to `t_el_forest_rps` and `t_forest_rsp`
+* Add argument `col_symbol_size` to `g_forest` to control the relative size of symbols used in the plot.
 
 # tern 0.6.4